+++ packages/cli/src/commands/deploy/index.ts
@@ -63,7 +63,7 @@ import validatePaths, {
   validateRootDirectory,
 } from '../../util/validate-paths';
 import { help } from '../help';
-import { deployCommand } from './command';
+import { deployCommand, deprecatedArchiveSplitTgz } from './command';
 import parseTarget from '../../util/parse-target';
 import { DeployTelemetryClient } from '../../util/telemetry/commands/deploy';
 import output from '../../output-manager';
@@ -102,6 +102,7 @@ export default async (client: Client): Promise<number> => {
     );
     telemetryClient.trackCliFlagPublic(parsedArguments.flags['--public']);
     telemetryClient.trackCliFlagLogs(parsedArguments.flags['--logs']);
+    telemetryClient.trackCliFlagNoLogs(parsedArguments.flags['--no-logs']);
     telemetryClient.trackCliFlagForce(parsedArguments.flags['--force']);
     telemetryClient.trackCliFlagWithCache(
       parsedArguments.flags['--with-cache']
@@ -111,6 +112,10 @@ export default async (client: Client): Promise<number> => {
       telemetryClient.trackCliFlagConfirm(parsedArguments.flags['--confirm']);
       output.warn('`--confirm` is deprecated, please use `--yes` instead');
       parsedArguments.flags['--yes'] = parsedArguments.flags['--confirm'];
+    }
+
+    if ('--logs' in parsedArguments.flags) {
+      output.warn('`--logs` is deprecated and now the default behavior.');
     }
   } catch (error) {
     printError(error);
@@ -215,10 +220,26 @@ export default async (client: Client): Promise<number> => {
     flags: parsedArguments.flags,
   });
 
-  const archive = parsedArguments.flags['--archive'];
-  if (typeof archive === 'string' && !isValidArchive(archive)) {
+  const parsedArchive = parsedArguments.flags['--archive'];
+  if (
+    typeof parsedArchive === 'string' &&
+    !(
+      isValidArchive(parsedArchive) ||
+      parsedArchive === deprecatedArchiveSplitTgz
+    )
+  ) {
     output.error(`Format must be one of: ${VALID_ARCHIVE_FORMATS.join(', ')}`);
     return 1;
+  }
+  if (parsedArchive === deprecatedArchiveSplitTgz) {
+    output.print(
+      `${prependEmoji(
+        `${param('--archive=tgz')} now has the same behavior as ${param(
+          '--archive=split-tgz'
+        )}. Please use ${param('--archive=tgz')} instead.`,
+        emoji('warning')
+      )}\n`
+    );
   }
 
   // Retrieve `project` and `org` from linked Project.
@@ -447,6 +468,7 @@ export default async (client: Client): Promise<number> => {
   const deployStamp = stamp();
   let deployment = null;
   const noWait = !!parsedArguments.flags['--no-wait'];
+  const withLogs = '--no-logs' in parsedArguments.flags ? false : true;
 
   const localConfigurationOverrides = pickOverrides(localConfig);
 
@@ -489,7 +511,7 @@ export default async (client: Client): Promise<number> => {
       target,
       skipAutoDetectionConfirmation: autoConfirm,
       noWait,
-      withLogs: parsedArguments.flags['--logs'],
+      withLogs,
       autoAssignCustomDomains,
     };
 
@@ -534,7 +556,7 @@ export default async (client: Client): Promise<number> => {
       createArgs,
       org,
       !project,
-      archive
+      parsedArchive ? 'tgz' : undefined
     );
 
     if (deployment instanceof NotDomainOwner) {
@@ -646,11 +668,13 @@ export default async (client: Client): Promise<number> => {
 
     if (err instanceof BuildError) {
       output.error(err.message || 'Build failed');
-      output.error(
-        `Check your logs at https://${now.url}/_logs or run ${getCommandName(
-          `logs ${now.url}`
+      output.log('\n');
+      output.log(
+        `To check build logs run: ${getCommandName(
+          `inspect ${now.url} --logs`
         )}`
       );
+      output.log(`Or inspect them in your browser at https://${now.url}/_logs`);
 
       return 1;
     }