+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @vercel/frameworks
 
+## 3.7.0
+
+### Minor Changes
+
+- Add Nitro to framework list ([#13472](https://github.com/vercel/vercel/pull/13472))
+
+## 3.6.4
+
+### Patch Changes
+
+- [frameworks] updated Docusaurus to match out framework naming standards ([#13333](https://github.com/vercel/vercel/pull/13333))
+
+## 3.6.3
+
+### Patch Changes
+
+- Rename SvelteKit (v1) ([#13255](https://github.com/vercel/vercel/pull/13255))
+
+## 3.6.2
+
+### Patch Changes
+
+- Replace Ember Mascot with Ember Framework Logo ([#13109](https://github.com/vercel/vercel/pull/13109))
+
+- Updates the [Astro](https://astro.build) logo ([#13112](https://github.com/vercel/vercel/pull/13112))
+
+## 3.6.1
+
+### Patch Changes
+
+- Enable "demo" property for "react-router" framework ([#13008](https://github.com/vercel/vercel/pull/13008))
+
+## 3.6.0
+
+### Minor Changes
+
+- Add "react-router" framework entry ([#12932](https://github.com/vercel/vercel/pull/12932))
+
+## 3.5.0
+
+### Minor Changes
+
+- Make vite detection supersede ionic-react ([#12880](https://github.com/vercel/vercel/pull/12880))
+
 ## 3.4.0
 
 ### Minor Changes